requirements.txt
streamlit
pymongo
python-dotenv
groq
requests